changed: move banner printing to separate compile unit
